requirements.txt
 openai
 voicevox-client
 pyaudio
 textwrap3
 simpleaudio
 asyncio
 translate